offload cache to separate module
