+++ tests/drivers/test_raster_drivers.py
@@ -1,4 +1,3 @@
-import zlib
 import pytest
 
 import rasterio
@@ -563,12 +562,3 @@ def test_compute_metadata_unoptimized(unoptimized_raster_file):
     )
 
     assert geometry_mismatch(shape(mtd['convex_hull']), convex_hull) < 1e-8
-
-
-def test_sizeof():
-    from terracotta.drivers.raster_base import _get_size_of
-    tile_shape = (256, 256)
-    data = zlib.compress(np.ones(tile_shape), 9)
-    mask = zlib.compress(np.zeros(tile_shape), 9)
-    size = _get_size_of((data, mask, 'float64', tile_shape))
-    assert 1450 < size < 1550
